Post

+++ Assets/Puzzle/Scripts/Firebase/Model/UserDto.cs
@@ -10,5 +10,5 @@ public class UserDto
     public string cedula;
     public string email;
     public string ciudad;
-    public int scoreSeg;
+    public float scoreSeg;
 }
